+++ packages/experimental/Badge/src/PresenceBadge/PresenceBadge.types.ts
@@ -3,12 +3,12 @@ import { BadgeCoreTokens } from '../Badge.types';
 
 export const presenceBadgeName = 'PresenceBadge';
 export const PresenceBadgeStatuses = ['doNotDisturb', 'busy', 'unknown', 'blocked', 'outOfOffice', 'away', 'available', 'offline'] as const;
-export type Presence = typeof PresenceBadgeStatuses[number];
+export type PresenceBadgeStatus = typeof PresenceBadgeStatuses[number];
 
 export interface PresenceBadgeTokens extends BadgeCoreTokens {}
 export interface PresenceBadgeProps extends BadgeCoreProps {
-  presence?: Presence;
-  isOutOfOffice?: boolean;
+  status?: PresenceBadgeStatus;
+  outOfOffice?: boolean;
 }
 
 export interface PresenceBadgeSlotProps extends Omit<BadgeSlotProps, 'text' | 'icon'> {}