+++ packages/experimental/Checkbox/src/index.ts
@@ -1,2 +1,12 @@
-export * from './Checkbox';
-export * from './Checkbox.types';
+export {
+  checkboxNameV1 as checkboxName,
+  CheckboxInfo,
+  CheckboxProps,
+  CheckboxSize,
+  CheckboxShape,
+  CheckboxSlotProps,
+  CheckboxState,
+  CheckboxTokens,
+  CheckboxType,
+} from '@fluentui-react-native/checkbox';
+export { Checkbox } from './Checkbox';